Conflicts:
	app/code/Magento/Ui/Component/Form/Element/AbstractOptionsField.php

+++ app/code/Magento/Downloadable/Model/Link/SaveHandler.php
@@ -7,11 +7,12 @@
 namespace Magento\Downloadable\Model\Link;
 
 use Magento\Downloadable\Api\LinkRepositoryInterface as LinkRepository;
+use Magento\Framework\EntityManager\Operation\ExtensionInterface;
 
 /**
  * Class SaveHandler
  */
-class SaveHandler
+class SaveHandler implements ExtensionInterface
 {
     /**
      * @var LinkRepository
@@ -29,22 +30,32 @@ class SaveHandler
     /**
      * @param string $entityType
      * @param object $entity
-     * @return object
+     * @param array $arguments
+     * @return \Magento\Catalog\Api\Data\ProductInterface|object
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function execute($entityType, $entity)
+    public function execute($entityType, $entity, $arguments = [])
     {
+        /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
         if ($entity->getTypeId() !== 'downloadable') {
             return $entity;
         }
-        /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
-        foreach ($this->linkRepository->getList($entity->getSku()) as $link) {
-            $this->linkRepository->delete($link->getId());
-        }
+
+        $oldLinks = $this->linkRepository->getList($entity->getSku());
         $links = $entity->getExtensionAttributes()->getDownloadableProductLinks() ?: [];
+        $updatedLinkIds = [];
         foreach ($links as $link) {
+            if ($link->getId()) {
+                $updatedLinkIds[] = $link->getId();
+            }
             $this->linkRepository->save($entity->getSku(), $link, !(bool)$entity->getStoreId());
         }
+        foreach ($oldLinks as $link) {
+            if (!in_array($link->getId(), $updatedLinkIds)) {
+                $this->linkRepository->delete($link->getId());
+            }
+        }
+
         return $entity;
     }
 }
